+++ src/test/kotlin/blackjack/domain/model/card/DeckTest.kt
@@ -1,9 +1,5 @@
-package blackjack.domain.model
+package blackjack.domain.model.card
 
-import blackjack.domain.model.card.Card
-import blackjack.domain.model.card.CardNumber
-import blackjack.domain.model.card.Deck
-import blackjack.domain.model.card.Suit
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
